refactor: clean testing code
